refactor[./]: изменить роутинг
На сервере роуты books сгруппированы в общий роут
На сервере исправлено распределение запросов между admins и auth
На фронтенде заменены соотвествующие пути api
Создана папка /docs и создан routes.md
