+++ TODO.md
@@ -49,10 +49,10 @@
 
 1. Вынести какие-то стили в globalStyles.scss
 
-1. Сгруппировать роутеры на сервере по предметным областям (auth, books, users, etc.)
-
-1. Пересмотреть роутинг / (auth) и /admin, перенести /register
-
 1. Создать в /docs в корне проекта и прописать документацию, в т.ч. по роутерам.
 
 1. Перенести фото .main_page.png из корня в /docs, а то жирно ему в корне сидеть :)
+
+1. Решить проблему с указанием расширения для файлов .ts (.js) при импортах, чтобы расширение можно было не указывать.
+
+1. Сделать хэширование логина на сервере (по аналогии с хэшированием пароля).
