+++ frontend/src/features/auth/api/auth.ts
@@ -2,7 +2,7 @@ import { validateResponse } from '../../../app/validateResponse';
 import { Admin, AdminSchema } from '../../../entities/admin/model/Admin';
 
 export function login(login: string, password: string): Promise<void> {
-    return fetch("/api/login", {
+    return fetch("/api/auth/login", {
         method: "POST",
         headers: {
             "Content-Type": "application/json",
@@ -14,7 +14,7 @@ export function login(login: string, password: string): Promise<void> {
 }
 
 export function logout(): Promise<void> {
-    return fetch("api/logout", {
+    return fetch("api/auth/logout", {
         method: "POST",
         headers: {
             "Content-Type": "application/json",
@@ -26,7 +26,7 @@ export function logout(): Promise<void> {
 }
 
 export function fetchMe(): Promise<Admin> {
-    return fetch("/api/admins/me")
+    return fetch("/api/auth/me")
         .then(validateResponse)
         .then((response) => response.json())
         .then((data) => AdminSchema.parse(data));
